services/simulator_service/requirements.txt
requests>=2.26.0
python-dotenv==1.2.2
